\name{tsms-package}
\alias{tsms-package}
\alias{tsms}
\docType{package}
\title{
  Modeling Univariate Time Series with Multiple Seasonality
}
\description{
  This package combines seasonality detection and best fitting via extended seasonal ARMA model. It tends to achieve satisfactory 1-step ahead prediction comparing to regular ARMA model, and a comparable n-step ahead prediction regarding to non-parametric seasonal models.
}
\details{
  The package is beta version. The core function is HSARMA.auto
}
\author{
Maintainer: Tianyang Xie <xie00039@umn.edu>
}
\references{
  T. Xie, J. Ding, "Time series modeling with multiple seasonality," in preparation.
}
\link{
  FSARMA.fit
  FSARMA.pred
  HSARMA.auto
}

