\name{tsms-package}
\alias{tsms-package}
\alias{tsms}
\docType{package}
\title{
  Modeling Univariate Time Series with Multiple Seasonality
}
\description{
  This package introduces a tool box for generalized ARMA model, as well as the MS modeling procedure for modeling univariate time series with multiple seasonality.
}
\details{
  The tool box of generalized ARMA model consists of core functions: 'garma.fit()', 'garma.pred()', 'garma.gen()'.
  The MS modeling procedure utilize the function: 'MS()'.
}
\author{
Maintainer: Tianyang Xie <xie00039@umn.edu>
Co-author: Jie Ding <dingj@umn.edu>
}
\references{
  T. Xie, J. Ding, "Forecasting with Multiple Seasonality ," https://arxiv.org/abs/2008.12340.
}


